refactor(notifications): decouple cards from Slack delivery
Move reusable notification card rendering and data behind generic naming while keeping Slack-specific message and delivery behavior isolated.
