+++ packages/back-end/package.json
@@ -9,12 +9,12 @@
     "build:emails": "mkdir -p dist/templates && cp -r src/templates/* dist/templates",
     "assemble-skills": "node scripts/assemble-agent-skills.mjs",
     "copy-skills": "rimraf -rf dist/agent/skills && mkdir -p dist/agent/skills && cp -r generated/agent-skills/. dist/agent/skills || echo 'Warning: No agent skills to copy'",
-    "build:slack-assets": "mkdir -p dist/services/slack/assets && cp -r src/services/slack/assets/* dist/services/slack/assets",
+    "build:notification-card-assets": "mkdir -p dist/services/notificationCards/assets && cp -r src/services/notificationCards/assets/* dist/services/notificationCards/assets",
     "build:clean": "rimraf -rf dist",
     "build:typescript": "tsc",
     "clean": "rimraf node_modules",
     "dev": "pnpm build:clean && pnpm assemble-skills && concurrently \"pnpm watch-compile\" \"pnpm watch-dev | pnpm pino-pretty\"",
-    "build": "pnpm build:clean && pnpm assemble-skills && pnpm build:typescript && pnpm build:emails && pnpm build:slack-assets && pnpm copy-skills",
+    "build": "pnpm build:clean && pnpm assemble-skills && pnpm build:typescript && pnpm build:emails && pnpm build:notification-card-assets && pnpm copy-skills",
     "start": "node dist/server.js",
     "start:with-tracing": "node --require ./dist/tracing.opentelemetry.js dist/server.js",
     "start:with-datadog": "node --require ./dist/tracing.datadog.js dist/server.js",
