+++ packages/back-end/src/services/notificationCards/assets/inter-latin-400-normal.woff
