+++ packages/back-end/src/services/notificationCards/assets/inter-latin-500-normal.woff
