+++ packages/back-end/src/services/notificationCards/assets/inter-latin-600-normal.woff
