+++ packages/back-end/src/services/notificationCards/assets/roboto-mono-latin-500-normal.woff
