+++ packages/shared/src/validators/event-webhook.ts
@@ -1,6 +1,7 @@
 import { z } from "zod";
 import { NotificationEventName } from "shared/types/events/base-types";
 import { zodNotificationEventNamesEnum } from "./events";
+import { experimentCardFormats } from "./notification-card";
 
 export const eventWebHookPayloadTypes = [
   "raw",
@@ -31,43 +32,6 @@ export const slackEventWebHookMetadata = z
     isEnterpriseInstall: z.boolean().optional(),
   })
   .strict();
-
-export const experimentCardFormats = ["none", "compact", "detailed"] as const;
-
-export const slackCardKinds = [
-  "started",
-  "significance",
-  "won",
-  "lost",
-  "stopped",
-  "warning",
-  "decisionShip", // Decision Framework recommends shipping (still running)
-  "decisionRollback", // Decision Framework recommends rolling back
-] as const;
-export type SlackCardKind = (typeof slackCardKinds)[number];
-
-// Experiment events that post a results CARD image (not a plain text message),
-// mapped to the card kind they announce. Any event not listed posts text only.
-// Source of truth for the settings UI (badging card vs text) and previews; keep
-// in sync with compactEventForNotification (back-end). experiment.stopped.*
-// resolve their real kind (won/lost/stopped) from the run's results at delivery
-// time, so "stopped" here is only a preview hint.
-export const SLACK_CARD_EVENT_KINDS: Record<string, SlackCardKind> = {
-  "experiment.started": "started",
-  "experiment.info.significance": "significance",
-  "experiment.decision.ship": "decisionShip",
-  "experiment.decision.rollback": "decisionRollback",
-  "experiment.warning": "warning",
-  "experiment.health.guardrailFailed": "warning",
-  "experiment.health.noData": "warning",
-  "experiment.health.queryFailed": "warning",
-  "experiment.stopped.shipped": "stopped",
-  "experiment.stopped.rolledback": "stopped",
-};
-
-export const slackCardKindForEvent = (
-  eventName: string,
-): SlackCardKind | undefined => SLACK_CARD_EVENT_KINDS[eventName];
 
 // Digest cadence. `custom` delivers like the others but signals the UI to
 // expose the full day/time editor.
