+++ packages/shared/src/validators/index.ts
@@ -42,6 +42,7 @@ export * from "./webhooks";
 export * from "./event-forwarder-config";
 export * from "./event-forwarder-access-test";
 export * from "./event-forwarder-status";
+export * from "./notification-card";
 export * from "./event-webhook";
 export * from "./feature-webhook-schemas";
 // Load watch before events - events imports base-types which imports validators, creating a cycle.
